Disable hardware acceleration
The current state of Firefox hardware acceleration on Linux with an
Nvidia GPU means you get massive performance dips while videos (e.g.
YouTube) are playing. This makes it very unpleasant to do anything else
while a video is playing.

So far, there appears to be no performance hit to not using it.